requirements.txt
beautifulsoup4==4.8.0
bs4==0.0.1
requests==2.22.0
urllib3==1.25.3
python-telegram-bot==12.8.0
